Define MARS data levels in separate file.
